Re-define taxon primary key field
The id for a taxon is not intended to be auto generated, hence the
replacement of AutoField to PositiveBigIntegerField. Even though the ids
has not been autogenerated during an import, this makes it more clear.

Also, an autogenerated id for facts is not necesssary since it is a
one-to-one relationship to taxon and therefore it makes sense to use the
taxon id as primary key and to remove the autogenerated id field.

+++ taxa/models.py
@@ -81,8 +81,8 @@ class TaxonQuerySet(models.QuerySet):
 
 
 class Taxon(models.Model):
-    id = models.AutoField(primary_key=True, editable=False)
-    parent_id = models.IntegerField(blank=True, null=True, editable=False)
+    id = models.PositiveBigIntegerField(primary_key=True, editable=False)
+    parent_id = models.PositiveBigIntegerField(blank=True, null=True, editable=False)
     slug = models.SlugField(max_length=255, editable=False)
     scientific_name = models.CharField(max_length=200)
     authority = models.CharField(max_length=200, blank=True, null=True)
